+++ ams/api/src/main/java/com/netease/arctic/ams/api/events/mixed/ExpireMixedSnapshotsResult.java
@@ -1,0 +1,61 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.netease.arctic.ams.api.events.mixed;
+
+import com.netease.arctic.ams.api.events.ExpireResult;
+import org.apache.iceberg.relocated.com.google.common.collect.Sets;
+import org.immutables.value.Value;
+
+import java.util.Set;
+
+/** Event details for expiring snapshots in mixed format */
+@Value.Immutable
+public abstract class ExpireMixedSnapshotsResult implements ExpireResult {
+  @Value.Default
+  public Set<String> deletedBaseFiles() {
+    return Sets.newHashSet();
+  }
+
+  @Value.Default
+  public Set<String> deletedInsertFiles() {
+    return Sets.newHashSet();
+  }
+
+  @Value.Default
+  public Set<String> deletedPositionDeleteFiles() {
+    return Sets.newHashSet();
+  }
+
+  @Value.Default
+  public Set<String> deletedEqualityDeleteFiles() {
+    return Sets.newHashSet();
+  }
+
+  @Value.Default
+  public Set<String> deletedIcebergEqualityDeleteFiles() {
+    return Sets.newHashSet();
+  }
+
+  public abstract Set<String> deletedMetadataFiles();
+
+  @Value.Default
+  public Set<String> deletedStatisticsFiles() {
+    return Sets.newHashSet();
+  }
+}
