+++ ams/server/src/test/java/com/netease/arctic/server/optimizing/maintainer/TestSnapshotExpire.java
@@ -25,9 +25,9 @@ import com.google.common.collect.Lists;
 import com.netease.arctic.BasicTableTestHelper;
 import com.netease.arctic.TableTestHelper;
 import com.netease.arctic.ams.api.TableFormat;
-import com.netease.arctic.ams.api.events.expire.ExpireEvent;
-import com.netease.arctic.ams.api.events.expire.mixed.ExpireMixedSnapshotsResult;
-import com.netease.arctic.ams.api.events.expire.mixed.ImmutableExpireMixedSnapshotsResult;
+import com.netease.arctic.ams.api.events.ExpireEvent;
+import com.netease.arctic.ams.api.events.mixed.ExpireMixedSnapshotsResult;
+import com.netease.arctic.ams.api.events.mixed.ImmutableExpireMixedSnapshotsResult;
 import com.netease.arctic.catalog.BasicCatalogTestHelper;
 import com.netease.arctic.catalog.CatalogTestHelper;
 import com.netease.arctic.data.ChangeAction;
