+++ ams/server/src/test/java/com/netease/arctic/server/optimizing/maintainer/TestSnapshotExpireIceberg.java
@@ -21,8 +21,8 @@ package com.netease.arctic.server.optimizing.maintainer;
 import com.netease.arctic.BasicTableTestHelper;
 import com.netease.arctic.TableTestHelper;
 import com.netease.arctic.ams.api.TableFormat;
-import com.netease.arctic.ams.api.events.expire.iceberg.ExpireSnapshotsResult;
-import com.netease.arctic.ams.api.events.expire.iceberg.ImmutableExpireSnapshotsResult;
+import com.netease.arctic.ams.api.events.iceberg.ExpireSnapshotsResult;
+import com.netease.arctic.ams.api.events.iceberg.ImmutableExpireSnapshotsResult;
 import com.netease.arctic.catalog.BasicCatalogTestHelper;
 import com.netease.arctic.catalog.CatalogTestHelper;
 import org.apache.iceberg.Table;
